Simplify and harden S3 part recovery flow
